SVCS-790 restructure auth service
