+++ src/services/auth/oidc/index.ts
@@ -1,6 +1,6 @@
-import { HttpInstance, OidcService } from '../../types';
-import { HttpClient } from '../http-client';
-import { OidcLinkRequestBody } from './oidcTypes';
+import { HttpInstance } from '../../../types';
+import { HttpClient } from '../../http-client';
+import { OidcLinkRequestBody, OidcService } from './types';
 
 export default (
   oidcClient: HttpClient,
